requirements.txt

pyyaml
numpy
# Ajouter d'autres dépendances selon les modules utilisés
spidev
